requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data loading
tfrecord>=1.14.0

# Visualization
torchviz>=0.0.2
graphviz>=0.20
matplotlib>=3.7.0
onnx>=1.15.0
netron>=7.0.0

# Optional but recommended
scipy>=1.10.0
tqdm>=4.65.0

